+++ Tests/StreamChatTests/Database/DTOs/MemberModelDTO_Tests.swift
@@ -143,4 +143,64 @@ final class MemberModelDTO_Tests: XCTestCase {
         let loadedMember = try XCTUnwrap(database.viewContext.member(userId: userId, cid: cid))
         XCTAssertTrue(loadedQuery.members.contains(loadedMember))
     }
+
+    func test_saveMembers_whenFirstPage_clearPreviousMembersFromQuery() throws {
+        let userId: UserId = .unique
+        let cid: ChannelId = .unique
+        let members: ChannelMemberListPayload = .init(members: [.dummy(), .dummy()])
+        let query = ChannelMemberListQuery(cid: cid, filter: .equal(.isModerator, to: true))
+
+        // Save previous members
+        let previousMembers = try saveDummyMembers(toQuery: query, cid: cid)
+        XCTAssertEqual(previousMembers.count, 4)
+
+        // Save new members
+        var newMembers: [ChatChannelMember] = []
+        try database.writeSynchronously { session in
+            newMembers = try session.saveMembers(payload: members, channelId: cid, query: query)
+                .map { try $0.asModel() }
+        }
+
+        // Assert the members in the DB are only the new members.
+        let loadedQuery = try XCTUnwrap(database.viewContext.channelMemberListQuery(queryHash: query.queryHash))
+        XCTAssertEqual(Set(loadedQuery.members.map(\.user.id)), Set(newMembers.map(\.id)))
+    }
+
+    func test_saveMembers_whenAnotherPage_doesNotClearPreviousMembersFromQuery() throws {
+        let userId: UserId = .unique
+        let cid: ChannelId = .unique
+        let members: ChannelMemberListPayload = .init(members: [.dummy(), .dummy()])
+        var query = ChannelMemberListQuery(cid: cid)
+        query.pagination = .init(pageSize: 20, offset: 25)
+
+        // Save previous members
+        let previousMembers = try saveDummyMembers(toQuery: query, cid: cid)
+        XCTAssertEqual(previousMembers.count, 4)
+
+        // Save new members
+        var newMembers: [ChatChannelMember] = []
+        try database.writeSynchronously { session in
+            newMembers = try session.saveMembers(payload: members, channelId: cid, query: query)
+                .map { try $0.asModel() }
+        }
+
+        // Assert the members in the DB contain the old and new members
+        let loadedQuery = try XCTUnwrap(database.viewContext.channelMemberListQuery(queryHash: query.queryHash))
+        let allMembers = previousMembers + newMembers
+        XCTAssertEqual(Set(loadedQuery.members.map(\.user.id)), Set(allMembers.map(\.id)))
+    }
+
+    private func saveDummyMembers(
+        _ members: [MemberPayload] = [.dummy(), .dummy(), .dummy(), .dummy()],
+        toQuery query: ChannelMemberListQuery,
+        cid: ChannelId
+    ) throws -> [ChatChannelMember] {
+        let members: ChannelMemberListPayload = .init(members: [.dummy(), .dummy(), .dummy(), .dummy()])
+        try database.writeSynchronously { session in
+            try session.saveChannel(payload: self.dummyPayload(with: cid))
+            session.saveMembers(payload: members, channelId: cid, query: query)
+        }
+        let loadedQuery = try XCTUnwrap(database.viewContext.channelMemberListQuery(queryHash: query.queryHash))
+        return try loadedQuery.members.map { try $0.asModel() }
+    }
 }